+++ pkg/clients/authn.go
@@ -7,7 +7,6 @@ import (
 	"github.com/cyberark/conjur-api-go/conjurapi"
 	"github.com/cyberark/conjur-api-go/conjurapi/authn"
 	"github.com/cyberark/conjur-cli-go/pkg/prompts"
-	"github.com/manifoldco/promptui"
 )
 
 type oidcConjurClient interface {
@@ -15,8 +14,8 @@ type oidcConjurClient interface {
 }
 
 // Login attempts to login to Conjur and prompts the user for credentials interactively
-func Login(conjurClient ConjurClient, decoratePrompt func(prompt *promptui.Prompt) *promptui.Prompt) (ConjurClient, error) {
-	authenticatePair, err := LoginWithPromptFallback(decoratePrompt, conjurClient, "", "")
+func Login(conjurClient ConjurClient) (ConjurClient, error) {
+	authenticatePair, err := LoginWithPromptFallback(conjurClient, "", "")
 	if err != nil {
 		return nil, err
 	}
@@ -28,12 +27,11 @@ func Login(conjurClient ConjurClient, decoratePrompt func(prompt *promptui.Promp
 // If either the username or password is missing then a prompt is presented to interactively request
 // the missing information from the user
 func LoginWithPromptFallback(
-	decoratePrompt prompts.DecoratePromptFunc,
 	client ConjurClient,
 	username string,
 	password string,
 ) (*authn.LoginPair, error) {
-	username, password, err := prompts.MaybeAskForCredentials(decoratePrompt, username, password)
+	username, password, err := prompts.MaybeAskForCredentials(username, password)
 	if err != nil {
 		return nil, err
 	}